+++ src/types/interface.d.ts
@@ -3,9 +3,22 @@ import { LocationQueryRaw, RouteRecordName } from 'vue-router';
 
 import STYLE_CONFIG from '@/config/style';
 
+export interface RouteMeta {
+  title?: string | Record<string, string>;
+  icon?: string;
+  expanded?: boolean;
+  orderNo?: number;
+  hidden?: boolean;
+  hiddenBreadcrumb?: boolean;
+  single?: boolean;
+  keepAlive?: boolean;
+  frameSrc?: string;
+  frameBlank?: boolean;
+}
+
 export interface MenuRoute {
   path: string;
-  title?: string;
+  title?: string | Record<string, string>;
   name?: string;
   icon?:
     | string
@@ -14,7 +27,7 @@ export interface MenuRoute {
       };
   redirect?: string;
   children: MenuRoute[];
-  meta: any;
+  meta: RouteMeta;
 }
 
 export type ModeType = 'dark' | 'light';
